requirements.txt
# External dependencies
python-dotenv>=1.0.0

